wasserstein implementation and structural similarity
